+++ src/points/mapshaper-dot-density.js
@@ -1,6 +1,8 @@
 import geom from '../geom/mapshaper-geom';
 import { testPointInPolygon } from '../geom/mapshaper-polygon-geom';
 import { error } from '../utils/mapshaper-logging';
+import { MaxHeap } from '../simplify/mapshaper-heap';
+import utils from '../utils/mapshaper-utils';
 // TODO: optimize point-in-polygon tests for complex polygons and many points
 // import { PathIndex } from '../paths/mapshaper-path-index';
 
@@ -65,33 +67,43 @@ function placeDot(shp, arcs, grid, bounds) {
 }
 
 function DotGrid(bounds, approxQueries, evenness) {
-  var grid = [];
   var x0 = bounds.xmin;
   var y0 = bounds.ymin;
   var w = bounds.width();
   var h = bounds.height();
   var approxCells = approxQueries * 0.8;
-  var cols = Math.round(Math.sqrt(approxCells * w / h));
+  var cols = Math.round(Math.sqrt(approxCells * w / h)) || 1;
   var rows = Math.ceil(cols * h / w); // overshoots bbox height
+  var gridWidth = w;
+  var gridHeight = w * rows / cols;
   var cells = cols * rows;
-  var nextId = -1;
+  var cellId = -1;
+  var grid = initGrid(cells);
+  // data used by optimal method
   var bestPoints;
-  var probesBeforeRelaxation = Math.ceil(Math.pow(cells, 0.8));
-  var maxProbes = cells * 10;
+  var bestHeap;
+
+  // Estimate the initial distance threshold between dots (based on a square grid)
+  // When evenness == 1, the initial value should be larger than the final
+  //   spacing between dots, after all the dots are added.
+  // When evenness < 1 (dart-throwing mode) the distance threshold is reduced in
+  //   proportion to the value of evenness.
+  var k = Math.pow(evenness, 0.85); // applying a curve seems create a better scale
+  // From trial and error, 0.75 seems to give a good result.
+  var minDist = gridWidth / cols * 0.75 * k;
+
+  // metrics
   var queries = 0;
-
-  // set height to height of grid
-  h = w * rows / cols;
-
-  var k = Math.pow(evenness, 0.7); // seems to make a better transition
-  // Estimate of minimum distance between dots (based on a square grid)
-  // 0.7 seems about right from trial-and-error...
-  var minDist = w / cols * 0.7 * k;
+  var lastDistSq = 0;
+  var bestCount = 0;
 
   return {done: done, getPoint: getPoint};
 
   function done() {
-    // TODO: check placement metrics if debugging
+    // console.log( 'queries:', queries,'bestPct:', pct(bestCount, queries), "minDist:", Math.round(minDist), "lastDist:", Math.round(Math.sqrt(lastDistSq)));
+    function pct(a, b) {
+      return Math.round(a / b * 100) + '%';
+    }
   }
 
   function getPoint() {
@@ -101,22 +113,33 @@ function DotGrid(bounds, approxQueries, evenness) {
     return getSpacedPoint();
   }
 
+  function initGrid(n) {
+    var arr = [];
+    for (var i=0; i<n; i++) arr.push([]);
+    return arr;
+  }
+
   function getOptimalPoint() {
     var p;
-    // first, try to place a point in each grid cell
-    // (for fast and even initial dot placement)
-    while (++nextId < cells) {
-      p = getRandomPointInCell(nextId);
+    // first, try to place a random but well-spaced point in each grid cell
+    // (to create an initial sparse structure that gets filled in later)
+    while (++cellId < cells) {
+      p = getRandomPointInCell(cellId);
       if (pointIsUsable(p)) {
         return usePoint(p);
       }
     }
-    // assumes that the point grid has already been seeded
+    // fill in the gaps of the initial placement, starting with the largest gap
+    if (!bestPoints) {
+      initBestPoints();
+    }
     return useBestPoint();
   }
 
   function getSpacedPoint() {
     // use dart-throwing, reject points that are within the minimum distance
+    var probesBeforeRelaxation = Math.ceil(Math.pow(cells, 0.8));
+    var maxProbes = cells * 10;
     var probes = 0;
     var p;
     while (probes++ < maxProbes) {
@@ -135,83 +158,113 @@ function DotGrid(bounds, approxQueries, evenness) {
   // Add point to grid of used points
   function usePoint(p) {
     var i = pointToIdx(p);
-    var points = grid[i] || (grid[i] = []);
-    points.push(p);
+    grid[i].push(p);
     return p;
   }
 
   function useBestPoint() {
-    var maxDist = 0;
-    var bestId, p;
-    if (!bestPoints) {
-      bestPoints = initBestPoints();
-    }
-    for (var i=0; i<cells; i++) {
-      p = bestPoints[i];
-      if (p[2] > maxDist) {
-        maxDist = p[2];
-        bestId = i;
-      }
-    }
-    p = bestPoints[bestId].slice(0, 2); // remove distance member
+    var bestId = bestHeap.peek();
+    var p = bestPoints[bestId];
     usePoint(p); // add to grid of used points
-    updateBestPoints(p, bestId); // update best point of this cell and neighbors
+    updateNeighbors(p, bestId); // update best point of this cell and neighbors
+    lastDistSq = bestHeap.peekValue();
+    bestCount++;
     return p;
   }
 
   function initBestPoints() {
-    var bestPoints = [];
+    var values = [];
+    bestPoints = [];
+    var distSq;
     for (var i=0; i<cells; i++) {
-      bestPoints.push(findBestPointInCell(i));
+      distSq = findBestPointInCell(i);
+      values.push(distSq);
     }
-    return bestPoints;
+    bestHeap = new MaxHeap();
+    bestHeap.init(values);
   }
 
-  function updateBestPoints(p, i) {
+  function updateNeighbors(p, i) {
     var r = idxToRow(i);
     var c = idxToCol(i);
-    bestPoints[i] = findBestPointInCell(i);
-    updateBestPoint(p, c+1, r);
-    updateBestPoint(p, c, r+1);
-    updateBestPoint(p, c-1, r);
-    updateBestPoint(p, c, r-1);
-    updateBestPoint(p, c+1, r+1);
-    updateBestPoint(p, c-1, r+1);
-    updateBestPoint(p, c-1, r-1);
-    updateBestPoint(p, c+1, r-1);
+    updateBestPointInCell(i);
+    updateNeighbor(p, c+1, r);
+    updateNeighbor(p, c, r+1);
+    updateNeighbor(p, c-1, r);
+    updateNeighbor(p, c, r-1);
+    updateNeighbor(p, c+1, r+1);
+    updateNeighbor(p, c-1, r+1);
+    updateNeighbor(p, c-1, r-1);
+    updateNeighbor(p, c+1, r-1);
   }
 
-  function updateBestPoint(addedPt, c, r) {
+  function updateNeighbor(addedPt, c, r) {
     var i = colRowToIdx(c, r);
     if (i == -1) return;
     var bestPt = bestPoints[i];
     // don't need to update best point if the newly added point is too far away
-    // to have an effect
-    if (pointToPointDistSq(addedPt, bestPt) > bestPt[2]) return;
-    bestPoints[i] = findBestPointInCell(i);
+    // to have an effect.
+    // (about 80% of updates are skipped, typically)
+    if (distSq(addedPt, bestPt) < bestHeap.getValue(i)) {
+      updateBestPointInCell(i);
+    }
   }
 
-  function findBestPointInCell(i) {
-    var probes = 30;
+  function updateBestPointInCell(i) {
+    var distSq = findBestPointInCell(i);
+    bestHeap.updateValue(i, distSq);
+  }
+
+  function findBestPointInCell_v1(i) {
+    // randomly probe for the best point
+    var probes = 20;
     var maxDist = 0;
     var best, p, dist;
     while (probes-- > 0) {
       p = getRandomPointInCell(i);
-      dist = findDistanceFromNeighbors(p, i);
+      dist = findDistanceFromNeighbors(maxDist, p, i);
       if (dist > maxDist) {
         maxDist = dist;
         best = p;
       }
     }
-    best.push(maxDist); // add distance as third element in the point
-    return best;
+    bestPoints[i] = best;
+    return maxDist;
   }
 
-  function findDistanceFromNeighbors(xy, i) {
+  function findBestPointInCell(idx) {
+    // use a grid pattern to find the best point
+    var perSide = 5;
+    var maxDist = 0;
+    var best, p, dist;
+    for (var i=0, n=perSide*perSide; i<n; i++) {
+      p = getGridPointInCell(idx, i, perSide);
+      dist = findDistanceFromNeighbors(maxDist, p, idx);
+      if (dist > maxDist) {
+        maxDist = dist;
+        best = p;
+      }
+    }
+    bestPoints[idx] = best;
+    return maxDist;
+  }
+
+  function getGridPointInCell(cellIdx, i, n) {
+    var r = idxToRow(cellIdx);
+    var c = idxToCol(cellIdx);
+    var dx = (i % n + 0.5) / n;
+    var dy = (Math.floor(i / n) + 0.5) / n;
+    var x = (dx + c) / cols * w + x0;
+    var y = (dy + r) / rows * h + y0;
+    return [x, y];
+  }
+
+  function findDistanceFromNeighbors(memo, xy, i) {
     var dist = Infinity;
     var r = idxToRow(i);
     var c = idxToCol(i);
     dist = reduceDistance(dist, xy, c, r);
+    if (dist < memo) return 0; // 10-15% speedup
     dist = reduceDistance(dist, xy, c+1, r);
     dist = reduceDistance(dist, xy, c, r+1);
     dist = reduceDistance(dist, xy, c-1, r);
@@ -226,19 +279,20 @@ function DotGrid(bounds, approxQueries, evenness) {
   function reduceDistance(memo, xy, c, r) {
     var i = colRowToIdx(c, r);
     if (i == -1) return memo; // off the edge
-    var distSq = pointToPointsDistSq(xy, grid[i] || []);
-    return Math.min(memo, distSq);
+    var distSq = pointToPointsDistSq(xy, grid[i]);
+    return distSq < memo ? distSq : memo;
   }
 
   function pointToPointsDistSq(xy, points) {
-    var dist = Infinity;
+    var minDist = Infinity, dist;
     for (var i=0; i<points.length; i++) {
-      dist = Math.min(dist, pointToPointDistSq(xy, points[i]));
+      dist = distSq(xy, points[i]);
+      if (dist < minDist) minDist = dist;
     }
-    return dist;
+    return minDist;
   }
 
-  function pointToPointDistSq(a, b) {
+  function distSq(a, b) {
     var dx = a[0] - b[0];
     var dy = a[1] - b[1];
     return dx * dx + dy * dy;
@@ -280,14 +334,13 @@ function DotGrid(bounds, approxQueries, evenness) {
     var i = colRowToIdx(c, r);
     if (i == -1) return false;
     var points = grid[i];
-    if (!points || !testPointCollision(xy, points, minDist)) return false;
-    return true;
+    return testPointCollision(xy, points, minDist);
   }
 
   function testPointCollision(xy, points, dist) {
     var d2 = dist * dist;
     for (var i=0; i<points.length; i++) {
-      if (pointToPointDistSq(xy, points[i]) < d2) {
+      if (distSq(xy, points[i]) < d2) {
         return true;
       }
     }
@@ -323,7 +376,7 @@ function DotGrid(bounds, approxQueries, evenness) {
   }
 
   function idxToCol(i) {
-    return Math.floor(i % cols);
+    return i % cols;
   }
 
   function idxToRow(i) {