+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+v0.5.44
+* Added "per-dot=" option to the -dots command, for setting the value-to-dot ratio. To represent 100 people per dot on a population map, you would use per-dot=100.
+* Added "copy-fields=" option to -dots, to copy one or more data fields from the original polygon layer to the generated dot features.
+* Added "multipart" option to -dots, which combines groups of same-color dots into multi-part features.
+
 v0.5.43
 * Improved evenness of dot placement.
 * Replaced "random" and "spacing=" options with a single "evenness=" option, which varies from 0 (random placement) to 1 (very even).